requirements.txt
# Core ML
numpy>=1.26
pandas>=2.2
scikit-learn>=1.4
xgboost>=2.0
torch>=2.2          # for LSTM
shap>=0.45

# Experiment tracking
mlflow>=2.13

# API
fastapi>=0.111
uvicorn[standard]>=0.29
pydantic>=2.7
prometheus-client>=0.20

# Monitoring
evidently>=0.4

# Data
pyarrow>=15.0

# Config
pyyaml>=6.0

# Dev / testing
pytest>=8.0
pytest-cov>=5.0
httpx>=0.27        # needed by FastAPI TestClient
tqdm>=4.66

# Optional: Jupyter for notebooks
jupyter>=1.0
matplotlib>=3.8
seaborn>=0.13
